fix label construction for sentiment trees

+++ sst_main.py
@@ -168,16 +168,16 @@ def main():
         dev_loss, dev_pred = trainer.test(dev_dataset)
         test_loss, test_pred = trainer.test(test_dataset)
 
-        train_pearson = metrics.pearson(train_pred, train_dataset.labels)
-        train_mse = metrics.mse(train_pred, train_dataset.labels)
+        train_pearson = metrics.pearson(train_pred, train_dataset.labels.float())
+        train_mse = metrics.mse(train_pred, train_dataset.labels.float())
         logger.info('==> Epoch {}, Train \tLoss: {}\tPearson: {}\tMSE: {}'.format(
             epoch, train_loss, train_pearson, train_mse))
-        dev_pearson = metrics.pearson(dev_pred, dev_dataset.labels)
-        dev_mse = metrics.mse(dev_pred, dev_dataset.labels)
+        dev_pearson = metrics.pearson(dev_pred, dev_dataset.labels.float())
+        dev_mse = metrics.mse(dev_pred, dev_dataset.labels.float())
         logger.info('==> Epoch {}, Dev \tLoss: {}\tPearson: {}\tMSE: {}'.format(
             epoch, dev_loss, dev_pearson, dev_mse))
-        test_pearson = metrics.pearson(test_pred, test_dataset.labels)
-        test_mse = metrics.mse(test_pred, test_dataset.labels)
+        test_pearson = metrics.pearson(test_pred, test_dataset.labels.float())
+        test_mse = metrics.mse(test_pred, test_dataset.labels.float())
         logger.info('==> Epoch {}, Test \tLoss: {}\tPearson: {}\tMSE: {}'.format(
             epoch, test_loss, test_pearson, test_mse))
